techDebt(bulkUploadWorkersRefactor): Create workers folder in validate folder to separate worker validation logic
